Finally Confirmed working AXI
Got axi working using the axi_lite_master from Youtube Video. Confirmed using the ieee spi vhd block that it's able to communicate to SPI. Need to clean up the file interconnects, but it shold be ready to integrate. Also added the ability for the user to control # of DACs.