Added LDAC, CLRN. Fixed Dwell Timer.
Increased the time spent on sending out signals with SPI to improve readout on oscilloscope.

Added the new MODE and CTRL signals (Control is set up as [LDAC], [CLRN], [RSTN], [TRIGGER_PVP]). Mode 3 allows complete user control of the LDAC.

Added new axi_slv driver to allow for increased functionality; I was unsure if a direct 1 to 1 with a 32 bit slv would match, so this implementation made more logical sense. May need to change in future.